Refactor `proto::message::SignRequest` to module
