+++ src/proto/message.rs
@@ -1,6 +1,7 @@
 //! Agent protocol message structures.
 
 pub mod identity;
+pub mod sign;
 pub mod unparsed;
 
 use core::str::FromStr;
@@ -10,66 +11,13 @@ use ssh_key::{
     certificate::Certificate, private::KeypairData, public::KeyData, Algorithm, Error, Signature,
 };
 
-pub use self::{identity::*, unparsed::*};
+pub use self::{identity::*, sign::*, unparsed::*};
 use super::{
     extension::{KeyConstraintExtension, MessageExtension},
     PrivateKeyData, ProtoError,
 };
 
 type Result<T> = core::result::Result<T, ProtoError>;
-
-/// Signature request with data to be signed with a key in an agent.
-///
-/// This structure is sent in a [`Request::SignRequest`] (`SSH_AGENTC_SIGN_REQUEST`) message.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3.6](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.6)
-#[derive(Clone, PartialEq, Debug)]
-pub struct SignRequest {
-    /// The public key portion of the [`Identity`] in the agent to sign the data with
-    pub pubkey: KeyData,
-
-    /// Binary data to be signed
-    pub data: Vec<u8>,
-
-    /// Signature flags, as described in
-    /// [draft-miller-ssh-agent-14 § 3.6.1](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.6.1)
-    pub flags: u32,
-}
-
-impl Decode for SignRequest {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let pubkey = reader.read_prefixed(KeyData::decode)?;
-        let data = Vec::decode(reader)?;
-        let flags = u32::decode(reader)?;
-
-        Ok(Self {
-            pubkey,
-            data,
-            flags,
-        })
-    }
-}
-
-impl Encode for SignRequest {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        [
-            self.pubkey.encoded_len_prefixed()?,
-            self.data.encoded_len()?,
-            self.flags.encoded_len()?,
-        ]
-        .checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        self.pubkey.encode_prefixed(writer)?;
-        self.data.encode(writer)?;
-        self.flags.encode(writer)?;
-
-        Ok(())
-    }
-}
 
 /// A container for a public / private key pair, or a certificate / private key.
 ///
